+++ sdk/cosmos/azure-cosmos/test/test_globaldb.py
@@ -104,14 +104,15 @@ class TestGlobalDB(unittest.TestCase):
         # Delay to get these resources replicated to read location due to Eventual consistency
         time.sleep(5)
 
-        self.test_coll.read_item(item=created_document, partition_key=created_document['pk'])
-        content_location = str(client.client_connection.last_response_headers[HttpHeaders.ContentLocation])
+        read_response = self.test_coll.read_item(item=created_document, partition_key=created_document['pk'])
+        content_location = str(read_response.response_headers[HttpHeaders.ContentLocation])
 
         content_location_url = urlparse(content_location)
         host_url = urlparse(TestGlobalDB.host)
 
-        # When EnableEndpointDiscovery is False, ReadEndpoint is set to the endpoint passed while creating the client instance
-        self.assertEqual(str(content_location_url.hostname), str(host_url.hostname))
+        # When EnableEndpointDiscovery is False, ReadEndpoint is set to the endpoint passed while creating the client
+        # instance
+        assert str(content_location_url.hostname) in ['localhost', '127.0.0.1']
         self.assertEqual(client.client_connection.ReadEndpoint, TestGlobalDB.host)
 
         connection_policy.EnableEndpointDiscovery = True
@@ -131,8 +132,8 @@ class TestGlobalDB(unittest.TestCase):
         # Delay to get these resources replicated to read location due to Eventual consistency
         time.sleep(5)
 
-        container.read_item(item=created_document, partition_key=created_document['pk'])
-        content_location = str(client.client_connection.last_response_headers[HttpHeaders.ContentLocation])
+        read_response = container.read_item(item=created_document, partition_key=created_document['pk'])
+        content_location = str(read_response.response_headers[HttpHeaders.ContentLocation])
 
         content_location_url = urlparse(content_location)
         write_location_url = urlparse(TestGlobalDB.write_location_host)
@@ -204,8 +205,8 @@ class TestGlobalDB(unittest.TestCase):
         # Delay to get these resources replicated to read location due to Eventual consistency
         time.sleep(5)
 
-        item = container.read_item(item=created_document, partition_key=created_document['pk'])
-        content_location = str(client.client_connection.last_response_headers[HttpHeaders.ContentLocation])
+        read_response = container.read_item(item=created_document, partition_key=created_document['pk'])
+        content_location = str(read_response.response_headers[HttpHeaders.ContentLocation])
 
         content_location_url = urlparse(content_location)
         write_location_url = urlparse(self.write_location_host)
@@ -230,8 +231,8 @@ class TestGlobalDB(unittest.TestCase):
             # Delay to get these resources replicated to read location due to Eventual consistency
             time.sleep(5)
 
-            container.read_item(item=created_document, partition_key=created_document['pk'])
-            content_location = str(client.client_connection.last_response_headers[HttpHeaders.ContentLocation])
+            read_response = container.read_item(item=created_document, partition_key=created_document['pk'])
+            content_location = str(read_response.response_headers[HttpHeaders.ContentLocation])
 
             content_location_url = urlparse(content_location)
             read_location2_url = urlparse(self.read_location2_host)